+++ AdventOfCode2023/src/Day1.kt
@@ -11,8 +11,8 @@ class Day1 : IDailyPuzzle {
         return "/day1.txt"
     }
 
-    override fun getResultPuzzle1(input: String): Int {
-        return input.lineSequence().map { line -> (line.find { it.isDigit() }.toString() + line.findLast { it.isDigit() }).toInt() }.sum()
+    override fun getResultPuzzle1(input: String): Long {
+        return input.lineSequence().map { line -> (line.find { it.isDigit() }.toString() + line.findLast { it.isDigit() }).toLong() }.sum()
     }
 
     private val digitMap = mapOf(Pair("one", "1"),
@@ -25,10 +25,10 @@ class Day1 : IDailyPuzzle {
             Pair("eight", "8"),
             Pair("nine", "9"))
 
-    override fun getResultPuzzle2(input: String): Int {
+    override fun getResultPuzzle2(input: String): Long {
         val digitList = digitMap.keys.toMutableList()
         digitList.addAll(digitMap.values)
-        return input.lineSequence().map { line -> (literal2numeral(line.findAnyOf(digitList)!!.second) + literal2numeral(line.findLastAnyOf(digitList)!!.second)).toInt() }.sum()
+        return input.lineSequence().map { line -> (literal2numeral(line.findAnyOf(digitList)!!.second) + literal2numeral(line.findLastAnyOf(digitList)!!.second)).toLong() }.sum()
     }
 
     private fun literal2numeral(value: String): String {
